+++ src/types/referoProps.ts
@@ -1,5 +1,5 @@
 import { QuestionnaireResponse, Attachment, Questionnaire, QuestionnaireItem, QuestionnaireResponseItemAnswer, ValueSet } from 'fhir/r4';
-import { FieldValues, UseFormProps, UseFormReturn } from 'react-hook-form';
+import { UseFormProps, UseFormReturn } from 'react-hook-form';
 import { Store } from 'redux';
 
 import { MimeTypes, UploadFile } from '@helsenorge/file-upload/components/file-upload';
@@ -11,6 +11,8 @@ import { IQuestionnaireInspector } from '../util/questionnaireInspector';
 import { Resources } from '../util/resources';
 import { ValidationSummaryPlacement } from './formTypes/validationSummaryPlacement';
 import { TextMessage } from './text-message';
+
+import { DefaultValues } from '@/validation/defaultFormValues';
 
 export interface RenderCustomButtonsArgs {
   /**
@@ -67,7 +69,7 @@ export interface RenderCustomButtonsArgs {
    * `reactHookFormMethods.getValues`, etc., to implement custom button behaviors
    * that interact directly with the underlying form state.
    */
-  reactHookFormMethods: UseFormReturn<FieldValues, unknown>;
+  reactHookFormMethods: UseFormReturn<DefaultValues, unknown, DefaultValues>;
 }
 
 export interface ReferoProps {